Use a buffered writer to write to post files
This drastically reduces the number of write syscalls being made.

This is a very trivial change to make in pure zig. It was slightly more
involved than that in this case.

md4c takes a void* pointer to some "userdata" that it passes back into
the process_output callback. Seems only logical that you'd pass the
writer object this way, so that you can actually write to it inside the
callback.

Once you receive the pointer in the callback however, you'd need to know
the exact type of the writer in order to cast it back and use it. I've
tried using the std.fs.File.Writer type, but that didn't seem to work
(runtime crash. most likely because the buffered writer object has a
different in-memory layout from an std.fs.File.Writer.)

@TypeOf() can be used to infer the type of the given writer, but this
information is comptime only and can't be passed to the callback at
runtime.

The solution? Create the callback function as an anonymous function
(using the struct member function trick) *inside* the parse function,
which has access to the @TypeOf() information. Now, it's simply a matter
of using this information inside the callback function and Bob's your
uncle :D